The overall logic is working, there are issues in search and flow the shabad is printed on screen. Discuss it with Harjas
